% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{make_plpr_data}
\alias{make_plpr_data}
\title{Simulated Data Frame}
\usage{
make_plpr_data(
  n_obs = 500,
  t_per = 10,
  dim_x = 20,
  theta = 0.5,
  pi = 0.8,
  sigma_ur = 0.6,
  rho = 0.8
)
}
\arguments{
\item{n_obs}{(\code{integer(1)}) \cr
The number of cross-sectional observations (i) to simulate.}

\item{t_per}{(\code{integer(1)}) \cr
The number of time periods (t) to simulate.}

\item{dim_x}{(\code{integer(1)}) \cr
The number of covariates.}

\item{theta}{(\code{numeric(1)}) \cr
The value of the causal parameter.}

\item{pi}{(\code{numeric(1)}) \cr
The value of the first-stage parameter.}

\item{sigma_ur}{(\code{numeric(1)}) \cr
Correlation between error tems U and R. The parameter governs model endogeneity.}

\item{rho}{(\code{numeric(1)}) \cr
Parameter governing the relationship between the covariates and the unobserved
individual heterogeneity. The value is chosen between 0 (pure random effect)
and 1 (pure fixed effects).}
}
\value{
A data object.
}
\description{
Generates data from a partially linear panel regression (PLPR) model with
instrumental variables as in Baiardi et al. (2026).

The data generating process is defined as

\eqn{Y_{it} = \theta D_{it} + g_0(X_{it}) + \alpha_i + U_{it},}
\eqn{D_{it} = r_0(X_{it}) + \zeta_i + R_{it},}
\eqn{Z_{it} = m_0(X_{it}) + \gamma_i + V_{it},}
where \eqn{U_{it} \sim \mathcal{N}(0,1)}, \eqn{R_{it} \sim \mathcal{N}(0,1)},
\eqn{V_{it}  \sim \mathcal{N}(0,0.25^2)}, \eqn{\sigma_{ur} = 0.6},
\eqn{\alpha_i = \rho A_i + \sqrt{1-\rho^2} B_i} with
\eqn{Gamma_i\sim \mathcal{N}(3,3)}, \eqn{A_i\sim \mathcal{N}(0,1)}, and  \eqn{B_i\sim \mathcal{N}(0,5)}.

The covariates are distributed as \eqn{X_{it,p} \sim A_i + \mathcal{N}(0, 5)},
where \eqn{p} is the number of covariates.

The nuisance functions are given by

m_0 = 0.5 * X_1 + 0.5 *  X_1 * as.numeric(X_1>0) + 0.5 *  X_3
r_0 = 0.5 * X_1 + 0.5 *  X_1 * as.numeric(X_1>0) + 0.5 *  X_3
g_0 = 0.5 * X_1 + 0.5 *  X_1 * as.numeric(X_1>0) + 0.5 *  X_3
}
\examples{
df = make_plpr_data(n_obs = 500, t_per = 10, dim_x = 20,
                    theta = 0.5, pi = 0.8,
                    sigma_ur = 0.6, rho=0.8)

}
