% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtivdml_data.R
\name{xtivdml_data}
\alias{xtivdml_data}
\title{Data-backed DML Environment}
\description{
Data-backed environment for Double machine learning (DML) that cannot be initialized.
\code{xtivdml_data} sets up the data environment for panel data analysis with transformed variables.
The \code{\link[=xtivdml_data_from_data_frame]{xtivdml_data_from_data_frame()}} function can be used to create a new
instance of \code{xtivdml_data} from a \code{data.frame}.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{all_variables}}{(\code{character()})\cr
All variables available in the dataset.}

\item{\code{d_cols}}{(\code{character()})\cr
The treatment variable(s).}

\item{\code{data}}{(\code{\link[data.table:data.table]{data.table}})\cr
Data object.}

\item{\code{data_model}}{(\code{\link[data.table:data.table]{data.table}})\cr
Internal data object that implements the causal model as specified by
the user via \code{y_col}, \code{d_cols}, \code{x_cols},  and \code{z_cols}.}

\item{\code{n_instr}}{(\code{NULL}, \code{integer(1)}) \cr
The number of instruments.}

\item{\code{n_obs}}{(\code{integer(1)}) \cr
The number of observations.}

\item{\code{n_treat}}{(\code{integer(1)}) \cr
The number of treatment variables.}

\item{\code{treat_col}}{(\code{character(1)}) \cr
"Active" treatment variable in the multiple-treatment case.}

\item{\code{x_cols}}{(\code{NULL}, \code{character()}) \cr
The covariates. If \code{NULL}, all variables (columns of \code{data}) which are
neither specified as outcome variable \code{y_col}, nor as treatment variables
\code{d_cols}, nor as instrumental variables \code{z_cols} are used as covariates.
Default is \code{NULL}.}

\item{\code{y_col}}{(\code{character(1)}) \cr
The outcome variable.}

\item{\code{z_cols}}{(\code{NULL}, \code{character()}) \cr
The instrumental variables. Default is \code{NULL}.}

\item{\code{cluster_cols}}{(\code{character()})\cr
The cluster variable(s).}

\item{\code{n_cluster_vars}}{(\code{integer(1)}) \cr
The number of cluster variables.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-xtivdml_data-new}{\code{xtivdml_data$new()}}
\item \href{#method-xtivdml_data-print}{\code{xtivdml_data$print()}}
\item \href{#method-xtivdml_data-set_data_model}{\code{xtivdml_data$set_data_model()}}
\item \href{#method-xtivdml_data-clone}{\code{xtivdml_data$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-xtivdml_data-new"></a>}}
\if{latex}{\out{\hypertarget{method-xtivdml_data-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{xtivdml_data$new(
  data = NULL,
  x_cols = NULL,
  y_col = NULL,
  d_cols = NULL,
  z_cols = NULL,
  cluster_cols = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{(\code{\link[data.table:data.table]{data.table}}, \code{data.frame()})\cr
Data object.}

\item{\code{x_cols}}{(\code{character()}) \cr}

\item{\code{y_col}}{(\code{character(1)}) \cr
The outcome variable.}

\item{\code{d_cols}}{(\code{character(1)}) \cr
The treatment variable.}

\item{\code{z_cols}}{(\code{NULL}, \code{character()}) \cr
The instrumental variables. Default is \code{NULL}.}

\item{\code{cluster_cols}}{(\code{character()}) \cr
The cluster variable(s).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-xtivdml_data-print"></a>}}
\if{latex}{\out{\hypertarget{method-xtivdml_data-print}{}}}
\subsection{Method \code{print()}}{
Print \code{xtivdml_data} objects.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{xtivdml_data$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-xtivdml_data-set_data_model"></a>}}
\if{latex}{\out{\hypertarget{method-xtivdml_data-set_data_model}{}}}
\subsection{Method \code{set_data_model()}}{
Setter function for \code{data_model}. The function implements the causal model
as specified by the user via \code{y_col}, \code{d_cols}, \code{x_cols}, \code{z_cols} and
\code{cluster_cols} and assigns the role for the treatment variables in the
multiple-treatment case.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{xtivdml_data$set_data_model(treatment_var)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{treatment_var}}{(\code{character()})\cr
Active treatment variable that will be set to \code{treat_col}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-xtivdml_data-clone"></a>}}
\if{latex}{\out{\hypertarget{method-xtivdml_data-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{xtivdml_data$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
