% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtivdml_data.R
\name{xtivdml_data_from_data_frame}
\alias{xtivdml_data_from_data_frame}
\title{Wrapper for Double machine learning data-backend initialization from
data.frame.}
\usage{
xtivdml_data_from_data_frame(
  df,
  x_cols = NULL,
  y_col = NULL,
  d_cols = NULL,
  z_cols = NULL,
  cluster_cols = NULL
)
}
\arguments{
\item{df}{(\code{data.frame()})\cr
Data object.}

\item{x_cols}{(\code{character()}) \cr
The covariates.}

\item{y_col}{(\code{character(1)}) \cr
The outcome variable.}

\item{d_cols}{(\code{character()}) \cr
The treatment variable(s).}

\item{z_cols}{(\code{NULL}, \code{character()}) \cr
The instrumental variables. Default is \code{NULL}.}

\item{cluster_cols}{(\code{NULL}, \code{character()}) \cr
The cluster variables. Default is \code{NULL}.}
}
\value{
Creates a new instance of class \code{xtivdml_data}.
}
\description{
Initalization of xtivdml_data from \code{data.frame}.
}
\examples{
# Generate simulated panel dataset from `xtivdml`
data = make_plpr_data(n_obs = 500, t_per = 10, dim_x = 30, theta = 0.5, pi = 0.8,
                    sigma_ur = 0.6, rho=0.8)

# Set up DML data environment
x_cols  = paste0("X", 1:30)

obj_xtivdml_data = xtivdml_data_from_data_frame(data,
                x_cols = x_cols,  y_col = "y",
                d_cols = "d", z_cols = "z",
                cluster_cols = "id")
obj_xtivdml_data$print()

}
